feat: tcp流的分层架构
